requirements.txt
django==4.2
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-tinymce
django-dotenv
djlint
django-ratelimit

daphne
channels
channels-redis

whitenoise
redis
celery 

requests
rav